Remove margins from anchor elements
There is a bug in Outlook 2016 (and possibly other versions) that leads
to a `p` element losing its `margin` if there is an `a` element inside
the `p`, and the `a` element has css `margin` specified. `a` elements
are by default inline and without margins, so it shouldn't cause any
problems to solved the issue by just not specifying `margin` for `a`
elements.