Buttons in IBM (Lotus) Notes
IBM Notes 9 (earlier Lotus Notes) requires the background color to be
set on the `a` element of buttons. In addition to fixing this, I've done
a bunch of cleanup. The file has gotten simpler and shorter, but there
should be no changes to how the buttons look or work. (Apart from that
they now work in IBM Notes 9 as well.) Just write a comment if you are
unsure about any of the changes.