Fix button width in Outlook's "View in Browser"
The grid rule

td.columns,
td.column,
th.columns,
th.column {
table:not(.button) {
width: 100%;
}
}

sets the inner table of buttons to `width: 100%`. This breaks the layout
when reading the email in IE by using the "View in Browser" feature in
Outlook. The wide inner table causes the buttons to break over multiple
lines. This fix sets the inner table to `width: auto` for non-expanded
buttons.